requirements.txt
# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0

# Medical imaging I/O
nibabel>=5.0.0

# Numerical and scientific computing
numpy>=1.24.0

# Configuration management
pyyaml>=6.0

# Progress bars for training
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0

# Optional but recommended for data handling
scipy>=1.10.0